feat: ValidateInput을 삭제하고 그 메서드들을 InputView, Lotto, LottoMachine에 나눠 옮김
